Ajout des query params + correction de bugs sur les controlleurs et amélioration de la qualité du code

+++ API/controllers/circuitController.js
@@ -3,7 +3,20 @@ const db = require('../models');
 const utils = require('./utils');
 module.exports = 
 {
-    
+    circuit : (req,res,next) => 
+    {
+        let id_user = utils.verifToken(req.headers['authorization']);
+        if(id_user)
+        {
+            db.Circuit.findByPk(req.params.id_circuit)
+            .then((circuit) => res.json(circuit))
+            .catch((err) => {if(err) res.sendStatus(500)})
+        }
+        else
+            res.sendStatus(401);
+    },
+
+    //////////////////////////////////////////////////////////
 
     circuits : (req,res,next) => 
     {
@@ -15,7 +28,7 @@ module.exports =
             .catch((err) => {if(err) res.sendStatus(500)})
         }
         else
-            res.sendStatus(401)
+            res.sendStatus(401);
     },
 
     //////////////////////////////////////////////////////////
@@ -35,6 +48,7 @@ module.exports =
                 published : 0,
                 version : 0,
                 level : req.body.level,
+                id_user : id_user
             })
             .then((circuit) =>
             {
@@ -46,35 +60,57 @@ module.exports =
             .catch((err) => {if(err) res.sendStatus(500)})
         }
         else
-            res.sendStatus(401)
+            res.sendStatus(401);
     },
 
     //////////////////////////////////////////////////////////
 
     publicationCircuit : (req,res,next) =>
     {
-        db.Circuit.findOne({where : {id_circuit : req.body.id_circuit}})
-        .then(circuit =>
+        let id_user = utils.verifToken(req.headers['authorization']);
+        if(id_user)
         {
-            circuit.update({published: !circuit.published})
-        })
-        .then(res.send(200))
+            db.Circuit.findOne({where : {id_circuit : req.body.id_circuit}})
+            .then(circuit =>
+            {
+                circuit.update({published: !circuit.published})
+            })
+            .then(res.send(200))
+        }
+        else
+            res.sendStatus(401);
+       
     },
 
     //////////////////////////////////////////////////////////
 
     deleteCircuit : (req,res,next) =>
     {
-        db.Circuit.destroy({where : {id_circuit : req.body.id_circuit}})
-        .then(a => res.send(a))
-        .catch(res.send(404))
+        let id_user = utils.verifToken(req.headers['authorization']);
+        if(id_user)
+        {
+            db.Circuit.destroy({where : {id_circuit : req.body.id_circuit}})
+            .then(a => res.sendStatus(200))
+            .catch(res.sendStatus(500))
+        }
+        else
+            res.sendStatus(401); 
     },
 
     //////////////////////////////////////////////////////////
 
     publishedCircuits : (req,res,next) =>
     {
-        db.Circuit.findAll({where : {published : true}})
+        let id_user = utils.verifToken(req.headers['authorization']);
+        if(id_user)
+        {
+            db.Circuit.findAll({where : {published : true}})
+            .then(a => res.sendStatus(200))
+            .catch(res.sendStatus(500))
+        }
+        else
+            res.sendStatus(401);
+        
     }
 
 }
